Add new error check and test for corruption on inflate field
I've had a few feeds that would no longer replicate because the
inflate field was reset to 'zero' after a few feeds have been
authorized.

This check will create a new error 'Missing feed mappings' if
an entry has a clock that has a different number of feeds than
the mappings generated from the head's feedsMessage.