Use Object.assign to ensure opts isn't mutated
Some users may re-use the opts object passed as a parameter,
which results in surprising behaviour protocol errors after multiple
calls to .replicate() eg. random errors such as:

 Only 128 feeds currently supported.
 Remote sent invalid feed message
 Remote message is larger than 8MB (max allowed)

There was also an instance of the opts object in the constructor
being mutated as well, so added an Object.assign for that as well.